Refactor file structure and modify confetti count
