+++ module/confetti.js
@@ -1,7 +1,7 @@
 // confetti file: download at https://www.cssscript.com/confetti-falling-animation/
 
 var confetti = {
-    maxCount: 200,		//set max confetti count
+    maxCount: 4000,		//set max confetti count
     speed: 5,			//set the particle animation speed
     frameInterval: 15,	//the confetti animation frame interval in milliseconds
     alpha: 1.0,			//the alpha opacity of the confetti (between 0 and 1, where 1 is opaque and 0 is invisible)
